+++ app/src/main/java/com/example/myapplication/recipehelpers/FluidProduct.java
@@ -1,14 +1,14 @@
 package com.example.myapplication.recipehelpers;
 
 public class FluidProduct extends Product{
-    private final float temperature;
+    private final double temperature;
 
-    public FluidProduct(ComponentType type, String name, float amount, float probability, int ignoredByProductivity, float extraCountFraction, float temperature) {
+    public FluidProduct(ComponentType type, String name, double amount, double probability, int ignoredByProductivity, double extraCountFraction, double temperature) {
         super(type, name, amount, probability, ignoredByProductivity, extraCountFraction);
         this.temperature = temperature;
     }
 
-    public float getTemperature() {
+    public double getTemperature() {
         return temperature;
     }
 }
